expose method to create dynamic partial update request from an other serializer class that the one in the service
