+++ django_socio_grpc/mixins.py
@@ -1,6 +1,5 @@
 from asgiref.sync import sync_to_async
 from google.protobuf import empty_pb2
-from rest_framework import serializers
 
 from django_socio_grpc.protobuf.generation_plugin import (
     FilterGenerationPlugin,
@@ -16,9 +15,12 @@ from .grpc_actions.placeholders import (
     SelfSerializer,
     StrTemplatePlaceholder,
 )
-from .grpc_actions.utils import get_serializer_base_name
+from .grpc_actions.utils import (
+    get_partial_update_request_from_service,
+    get_serializer_base_name,
+)
 from .settings import grpc_settings
-from .utils.constants import DEFAULT_LIST_FIELD_NAME, PARTIAL_UPDATE_FIELD_NAME, REQUEST_SUFFIX
+from .utils.constants import DEFAULT_LIST_FIELD_NAME, REQUEST_SUFFIX
 
 
 ############################################################
@@ -249,41 +251,9 @@ class UpdateModelMixin(GRPCActionMixin):
         }
 
 
-def _get_partial_update_request(service):
-    serializer_class = service.get_serializer_class()
-
-    class PartialUpdateMetaClass(serializers.SerializerMetaclass):
-        """
-        This metaclass exists so we can set the PARTIAL_UPDATE_FIELD_NAME variable as an attribute name of PartialUpdateRequest.
-        This can be replaced by just declaring in PartialUpdateRequest:
-        _partial_update_fields = serializers.ListField(child=serializers.CharField())
-        but this would not be dynamic if a constant changes or if we want it to be configurable in settings in the future.
-        This metaclass should inherit from DRF SerializerMetaclass as serializer has it's own metaclass to add _declared_fields attribute
-        Using PartialUpdateRequest.setattr is not enough as _declared_fields is done in metaclass so all fields should be declared before
-        """
-
-        def __new__(cls, name, bases, attrs):
-            attrs[PARTIAL_UPDATE_FIELD_NAME] = serializers.ListField(
-                child=serializers.CharField()
-            )
-            return super().__new__(cls, name, bases, attrs)
-
-    class PartialUpdateRequest(serializer_class, metaclass=PartialUpdateMetaClass):
-        class Meta(serializer_class.Meta): ...
-
-    # INFO - L.G. - 19/06/2022 - extra field needs to be appended to
-    # the serializer.
-    if (fields := getattr(PartialUpdateRequest.Meta, "fields", None)) and not isinstance(
-        fields, str
-    ):
-        PartialUpdateRequest.Meta.fields = (*fields, PARTIAL_UPDATE_FIELD_NAME)
-
-    return PartialUpdateRequest
-
-
 class PartialUpdateModelMixin(GRPCActionMixin):
     @grpc_action(
-        request=FnPlaceholder(_get_partial_update_request),
+        request=FnPlaceholder(get_partial_update_request_from_service),
         request_name=StrTemplatePlaceholder(
             f"{{}}PartialUpdate{REQUEST_SUFFIX}", get_serializer_base_name
         ),
